requirements.txt
discord.py==1.7.2
PyNaCl==1.4.0
requests==2.21.0
youtube-dl==2021.2.22